feat: initial the schema system index page
