+++ components/InputWithTemplate/index.ts
@@ -1,7 +1,5 @@
 import InputWithTemplate from "./InputWithTemplate.vue";
-import { Template } from "./types";
 
 export {
   InputWithTemplate,
-  Template
 }
