+++ pages/schema-system/index.vue
@@ -1,68 +1,111 @@
 <template>
-  <div
-    class="max-w-7xl mx-auto text-left border border-gray-200 overflow-hidden sm:rounded-md"
-  >
-    <div class="px-6 py-6">
-      <Steps :steps="steps" :current-step-id="currentStepId" />
-      <div class="flex-1 mt-10">
-        <div class="text-left mb-5">
-          <h1 class="text-xl font-bold">
-            {{ steps[state.stepIndex].title }}
-          </h1>
-          <p class="text-gray-500">
-            {{ steps[state.stepIndex].description }}
-          </p>
+  <div class="max-w-7xl mx-auto px-4">
+    <div class="lg:grid lg:grid-cols-12 lg:gap-8">
+      <div
+        class="sm:text-center md:max-w-2xl md:mx-auto lg:col-span-6 lg:text-left"
+      >
+        <h1>
+          <span
+            class="mt-1 block text-4xl tracking-tight font-extrabold sm:text-5xl xl:text-6xl"
+          >
+            <span class="block text-gray-900">Bytebase</span>
+            <span class="text-transparent bg-clip-text bg-gradient-to-br from-blue-400 to-indigo-700">
+              Schema System
+            </span>
+          </span>
+        </h1>
+        <h2
+          class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-xl lg:text-lg xl:text-xl"
+        >
+          Bytebase
+          <span
+            class="text-blue-600"
+            style="box-shadow: rgb(255, 255, 255) 0px -0.166667em 0px 0px inset, rgb(186, 230, 253) 0px -0.333333em 0px 0px inset;"
+          >
+            schema system
+          </span>
+          is the world's most popular open source database. Bytebase supports
+          MySQL database natively to manage schema changes, backups, record
+          migration history and etc.
+        </h2>
+        <div
+          class="mt-8 sm:max-w-lg sm:mx-auto sm:text-center lg:text-left lg:mx-0"
+        >
+          <ActionButton
+            :class-names="[
+              'text-white bg-indigo-600 hover:bg-indigo-700 focus:ring-indigo-700',
+              'md:px-8 md:py-4 md:text-2xl',
+            ]"
+          >
+            Create your schema guideline
+          </ActionButton>
+          <!-- <ActionSection :moduleName="'database'" /> -->
         </div>
-        <transition appear name="slide-from-right" mode="out-in">
-          <SchemaSystemEngine
-            v-if="currentStepId === 'Step 1'"
-            :engines="engines"
-            :active-engine-id="state.engine.id"
-            @select="(val) => (state.engine = val)"
-          />
-          <SchemaSystemConfig
-            v-else-if="currentStepId === 'Step 2'"
-            :rules="rules"
-            :select-rules="state.rules"
-            @select="onRuleSelect"
-            @remove="onRuleRemove"
-            @change="onRuleChange"
-          />
-          <SchemaSystemPreview
-            v-else-if="currentStepId === 'Step 3'"
-            id="preview"
-            :rules="state.rules"
-            :engine="state.engine"
-          />
-        </transition>
-        <div class="flex justify-end items-center mt-10">
-          <ActionButton
-            v-if="state.stepIndex > 0"
-            :class-names="['hover:bg-gray-200 focus:ring-gray-500', 'mr-5']"
-            @click="preStep"
+      </div>
+      <div
+        class="mt-12 relative sm:max-w-lg sm:mx-auto lg:mt-0 lg:max-w-none lg:mx-0 lg:col-span-6 lg:flex lg:items-center"
+      >
+        <div class="relative mx-auto w-full rounded-lg lg:max-w-md">
+          <img class="w-full" src="~/assets/illustration/instruction.webp" alt="" />
+        </div>
+      </div>
+    </div>
+    <div class="lg:grid lg:grid-cols-12 lg:gap-8">
+      <div
+        class="mt-12 relative sm:max-w-lg sm:mx-auto lg:mt-0 lg:max-w-none lg:mx-0 lg:col-span-6 lg:flex lg:items-center"
+      >
+        <div class="relative mx-auto w-full rounded-lg lg:max-w-md">
+          <img class="w-full" src="~/assets/bytebase_and_mysql.webp" alt="" />
+        </div>
+      </div>
+      <div
+        class="sm:text-center md:max-w-2xl md:mx-auto lg:col-span-6 lg:text-left"
+      >
+        <h1>
+          <span
+            class="mt-1 block text-4xl tracking-tight font-extrabold sm:text-5xl xl:text-6xl"
           >
-            Previous
-          </ActionButton>
-          <ActionButton
-            v-if="state.stepIndex < steps.length - 1"
-            :class-names="[
-              'text-white bg-indigo-600 hover:bg-indigo-700 focus:ring-indigo-700',
-            ]"
-            :disabled="nextStepDisabled"
-            @click="nextStep"
+            <span class="block text-gray-900">Bytebase</span>
+            <span class="text-transparent bg-clip-text bg-gradient-to-br from-blue-400 to-indigo-700">
+              Schema System
+            </span>
+          </span>
+        </h1>
+        <h2
+          class="mt-3 text-base text-gray-500 sm:mt-5 sm:text-xl lg:text-lg xl:text-xl"
+        >
+          Bytebase
+          <span
+            class="text-blue-600"
+            style="box-shadow: rgb(255, 255, 255) 0px -0.166667em 0px 0px inset, rgb(186, 230, 253) 0px -0.333333em 0px 0px inset;"
           >
-            Next: {{ steps[state.stepIndex + 1].title }}
-          </ActionButton>
-          <ActionButton
-            v-if="state.stepIndex === steps.length - 1"
-            :class-names="[
-              'text-white bg-indigo-600 hover:bg-indigo-700 focus:ring-indigo-700',
-            ]"
-            :disabled="!configuration"
-            @click="downConfiguration"
-          >
-            Download as image
-          </ActionButton>
+            schema system
+          </span>
+          is the world's most popular open source database. Bytebase supports
+          MySQL database natively to manage schema changes, backups, record
+          migration history and etc.
+        </h2>
+        <div
+          class="mt-8 sm:max-w-lg sm:mx-auto sm:text-center lg:text-left lg:mx-0"
+        >
+          <ActionSection :moduleName="'database'" />
+        </div>
+      </div>
+    </div>
+    <div class="flex mt-12 justify-center">
+      <div class="flex flex-row space-x-4">
+        <div class="flex mt-6 text-base tracking-tight text-gray-400">
+          More is coming
+        </div>
+        <div class="flex flex-row space-x-6">
+          <img class="h-12" src="~/assets/logo/db-postgres.png" alt="postgresql" />
+          <img class="h-12" src="~/assets/logo/db-tidb.png" alt="tidb" />
+          <img class="h-12" src="~/assets/logo/db-snowflake.png" alt="snowflake" />
+          <img
+            class="h-12"
+            src="~/assets/logo/db-clickhouse.png"
+            alt="clickhouse"
+          />
         </div>
       </div>
     </div>
@@ -70,218 +113,15 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, reactive } from "@nuxtjs/composition-api";
-// import { mapStores } from "pinia";
-import domtoimage from "dom-to-image";
-import Steps, { Step } from "../../components/Steps.vue";
+import { defineComponent } from "@nuxtjs/composition-api";
 import ActionButton from "../../components/ActionButton.vue";
-import SchemaSystemEngine from "../../components/SchemaSystem/SchemaSystemEngine.vue";
-import SchemaSystemPreview from "../../components/SchemaSystem/SchemaSystemPreview.vue";
-import SchemaSystemConfig from "../../components/SchemaSystem/SchemaSystemConfig.vue";
-// import { useNotificationStore } from "../store/notification";
-import {
-  Engine,
-  Rule,
-  RuleLevel,
-  SelectedRule,
-  SchemaConfiguration,
-  engines,
-  rules,
-  PayloadType,
-} from "../../common/schemaSystem";
-
-const steps: Step[] = [
-  {
-    id: "Step 1",
-    title: "Chose engine",
-    description: "Select a DB engine to build your schema system",
-  },
-  {
-    id: "Step 2",
-    title: "Configure rules",
-    description: "Select and configure rules for your schema system",
-  },
-  {
-    id: "Step 3",
-    title: "Download",
-    description: "Preview and download the configuration",
-  },
-];
-
-const validRule = (rule: SelectedRule): boolean => {
-  if (!rule.payload) {
-    return true;
-  }
-
-  for (const key of Object.keys(rule.payload)) {
-    const target = rule.payload[key].value;
-    // undefined value means user didn't use custom config.
-    // will use the default value for schema system.
-    if (target === undefined) {
-      continue;
-    }
-
-    switch (rule.payload[key].type) {
-      case PayloadType.String:
-        // string payload can use default value
-        continue;
-      case PayloadType.Template:
-        if (!target) {
-          return false;
-        }
-        continue;
-      case PayloadType.StringArray:
-        if (!Array.isArray(target) || !target.length) {
-          return false;
-        }
-        continue;
-      default:
-        return false;
-    }
-  }
-
-  return true;
-}
-
-interface LocalState {
-  stepIndex: number;
-  engine: Engine;
-  rules: SelectedRule[];
-}
 
 export default defineComponent({
   components: {
-    Steps,
     ActionButton,
-    SchemaSystemEngine,
-    SchemaSystemConfig,
-    SchemaSystemPreview,
   },
   setup() {
-    const state = reactive<LocalState>({
-      stepIndex: 0,
-      engine: engines[0],
-      rules: [],
-    });
-
-    return {
-      steps,
-      state,
-      rules,
-      engines,
-    };
+    
   },
-  computed: {
-    // ...mapStores(useNotificationStore),
-    currentStepId(): string {
-      return this.steps[this.state.stepIndex].id;
-    },
-    configuration(): SchemaConfiguration | undefined {
-      if (!this.state.engine) {
-        return;
-      }
-
-      return {
-        version: "0.0.1",
-        engine: {
-          [this.state.engine.id]: {
-            version: this.state.engine.version,
-            rules: this.state.rules.map((rule) => ({
-              id: rule.id,
-              level: rule.level,
-              payload: rule.payload
-                ? Object.keys(rule.payload).reduce((res, key) => {
-                    res[key] = rule.payload[key].value || rule.payload[key].default;
-                    return res;
-                  }, {} as { [key: string]: any })
-                : undefined,
-            })),
-          },
-        },
-      };
-    },
-    nextStepDisabled(): boolean {
-      switch (this.state.stepIndex) {
-        case 0:
-          return !this.state.engine;
-        case 1:
-          return !this.state.rules.length || !this.validRules;
-        default:
-          return false;
-      }
-    },
-    validRules(): boolean {
-      for (const rule of this.state.rules) {
-        if (!validRule(rule)) {
-          return false;
-        }
-      }
-      return true;
-    }
-  },
-  methods: {
-    nextStep() {
-      this.state.stepIndex = Math.min(
-        this.steps.length - 1,
-        this.state.stepIndex + 1
-      );
-    },
-    preStep() {
-      this.state.stepIndex = Math.max(0, this.state.stepIndex - 1);
-    },
-    onRuleSelect(rule: Rule) {
-      this.state.rules.push({
-        ...rule,
-        level: RuleLevel.Error,
-      });
-    },
-    onRuleRemove(rule: SelectedRule) {
-      const index = this.state.rules.findIndex((r) => r.id === rule.id);
-      this.state.rules = [
-        ...this.state.rules.slice(0, index),
-        ...this.state.rules.slice(index + 1),
-      ];
-    },
-    onRuleChange(rule: SelectedRule) {
-      const index = this.state.rules.findIndex((r) => r.id === rule.id);
-      this.state.rules = [
-        ...this.state.rules.slice(0, index),
-        rule,
-        ...this.state.rules.slice(index + 1),
-      ];
-    },
-    downConfiguration() {
-      const node = document.getElementById("preview");
-      if (!node) return;
-
-      domtoimage
-        .toPng(node, { bgcolor: "#fff", style: { padding: "1.5rem" } })
-        .then((dataUrl) => {
-          var link = document.createElement("a");
-          link.download = `${this.state.engine.name}-schema-system.png`;
-          link.href = dataUrl;
-          link.click();
-        });
-    },
-  },
-});
+})
 </script>
-
-<style scoped>
-.slide-from-right-enter-active {
-  transition: all 0.3s ease-in-out;
-}
-
-.slide-from-right-leave-active {
-  transition: all 0.3s ease-in-out;
-}
-
-.slide-from-right-enter {
-  transform: translateX(20px);
-  opacity: 0;
-}
-.slide-from-right-leave-to {
-  transform: translateX(-20px);
-  opacity: 0;
-}
-</style>
